Move out components
