+++ src/components/Inputs/RichTextEditor/MenuBar/MenuBar.tsx
@@ -1,36 +1,27 @@
 import { FC } from 'react';
 
-import {
-  code,
-  format_align_center,
-  format_align_left,
-  format_align_right,
-  format_bold,
-  format_clear,
-  format_italics,
-  format_list_bulleted,
-  format_list_numbered,
-  redo,
-  table_chart,
-  undo,
-} from '@equinor/eds-icons';
 import { tokens } from '@equinor/eds-tokens';
 import { Editor } from '@tiptap/react';
 
 import {
-  EditorPanel,
   OnImageUploadFn,
   RichTextEditorFeatures,
 } from '../RichTextEditor.types';
 import AddImage from './AddImage';
 import Links from './Links';
-import { Section } from './MenuBar.styles';
-import MenuButton from './MenuButton';
 import TableBar from './TableBar';
 import TextColor from './TextColor';
-import { amplify_h2, amplify_h3 } from 'src/components/Icons/AmplifyIcons';
 
 import styled from 'styled-components';
+import { History } from './History';
+import { Formating } from './Formating';
+import { Headers } from './Headers';
+import { Lists } from './Lists';
+import { Code } from './Code';
+import { Alignment } from './Alignment';
+import { Table } from './Table';
+import { ClearFormating } from './ClearFormating';
+import MenuButton from './MenuButton';
 
 const { colors, spacings } = tokens;
 
@@ -45,11 +36,6 @@ const Container = styled.div`
   gap: ${spacings.comfortable.small};
   padding: ${spacings.comfortable.small};
   border-bottom: 1px solid ${colors.ui.background__medium.rgba};
-`;
-
-const Divider = styled.hr`
-  width: 1px;
-  background: ${colors.ui.background__medium.rgba};
 `;
 
 interface MenuBarProps {
@@ -83,159 +69,20 @@ const MenuBar: FC<MenuBarProps> = ({ editor, features, onImageUpload }) => {
   );
 };
 
-const History = ({ editor, features }: EditorPanel) => {
-  const onUndo = () => editor.chain().focus().undo().run();
-  const onRedo = () => editor.chain().focus().redo().run();
-  if (!features.includes(RichTextEditorFeatures.HISTORY)) return;
-  return (
-    <Section>
-      <MenuButton
-        icon={undo}
-        onClick={onUndo}
-        disabled={!editor.can().undo()}
-      />
-      <MenuButton
-        icon={redo}
-        onClick={onRedo}
-        disabled={!editor.can().redo()}
-      />
-    </Section>
-  );
+export {
+  MenuBar,
+  History,
+  Formating,
+  Headers,
+  Lists,
+  TextColor,
+  Code,
+  Alignment,
+  Links,
+  AddImage,
+  Table,
+  ClearFormating,
+  MenuButton,
+  TableBar,
 };
-
-const Formating = ({ editor, features }: EditorPanel) => {
-  const toggleBold = () => editor.chain().focus().toggleBold().run();
-  const toggleItalic = () => editor.chain().focus().toggleItalic().run();
-  if (!features.includes(RichTextEditorFeatures.FORMATTING)) return;
-  return (
-    <Section>
-      <MenuButton
-        active={editor.isActive('bold')}
-        icon={format_bold}
-        onClick={toggleBold}
-      />
-      <MenuButton
-        active={editor.isActive('italic')}
-        icon={format_italics}
-        onClick={toggleItalic}
-      />
-    </Section>
-  );
-};
-
-const Headers = ({ editor, features }: EditorPanel) => {
-  const toggleH2 = () =>
-    editor.chain().focus().toggleHeading({ level: 2 }).run();
-  const toggleH3 = () =>
-    editor.chain().focus().toggleHeading({ level: 3 }).run();
-  if (!features.includes(RichTextEditorFeatures.HEADERS)) return;
-  return (
-    <Section>
-      <MenuButton
-        active={editor.isActive('heading', { level: 2 })}
-        icon={amplify_h2}
-        onClick={toggleH2}
-      />
-      <MenuButton
-        active={editor.isActive('heading', { level: 3 })}
-        icon={amplify_h3}
-        onClick={toggleH3}
-      />
-    </Section>
-  );
-};
-
-const Lists = ({ editor, features }: EditorPanel) => {
-  const toggleBulletList = () =>
-    editor.chain().focus().toggleBulletList().run();
-  const toggleOrderedList = () =>
-    editor.chain().focus().toggleOrderedList().run();
-  if (!features.includes(RichTextEditorFeatures.LISTS)) return;
-  return (
-    <Section>
-      <MenuButton
-        active={editor.isActive('bulletList')}
-        icon={format_list_bulleted}
-        onClick={toggleBulletList}
-      />
-      <MenuButton
-        active={editor.isActive('orderedList')}
-        icon={format_list_numbered}
-        onClick={toggleOrderedList}
-      />
-    </Section>
-  );
-};
-
-const Code = ({ editor, features }: EditorPanel) => {
-  const toggleCode = () => editor.chain().focus().toggleCodeBlock().run();
-  if (!features.includes(RichTextEditorFeatures.CODE)) return;
-  return (
-    <MenuButton
-      active={editor.isActive('codeBlock')}
-      icon={code}
-      onClick={toggleCode}
-    />
-  );
-};
-
-const Alignment = ({ editor, features }: EditorPanel) => {
-  const alignLeft = () => editor.chain().focus().setTextAlign('left').run();
-  const alignCenter = () => editor.chain().focus().setTextAlign('center').run();
-  const alignRight = () => editor.chain().focus().setTextAlign('right').run();
-  if (!features.includes(RichTextEditorFeatures.ALIGNMENT)) return;
-  return (
-    <Section>
-      <MenuButton
-        active={editor.isActive({ textAlign: 'left' })}
-        icon={format_align_left}
-        onClick={alignLeft}
-      />
-      <MenuButton
-        active={editor.isActive({ textAlign: 'center' })}
-        icon={format_align_center}
-        onClick={alignCenter}
-      />
-      <MenuButton
-        active={editor.isActive({ textAlign: 'right' })}
-        icon={format_align_right}
-        onClick={alignRight}
-      />
-    </Section>
-  );
-};
-
-const Table = ({ editor, features }: EditorPanel) => {
-  const createTable = () => {
-    editor
-      .chain()
-      .focus()
-      .insertTable({ rows: 1, cols: 3, withHeaderRow: false })
-      .run();
-  };
-  if (!features.includes(RichTextEditorFeatures.TABLE)) return;
-  if (editor.isActive('table')) return;
-  return <MenuButton icon={table_chart} onClick={createTable} />;
-};
-
-const ClearFormating = ({ editor, features }: EditorPanel) => {
-  const clearFormatting = () =>
-    editor.chain().focus().clearNodes().unsetAllMarks().run();
-  if (!features.includes(RichTextEditorFeatures.CLEAR_FORMATTING)) return;
-  return (
-    <>
-      <Divider />
-      <MenuButton
-        icon={format_clear}
-        onClick={clearFormatting}
-        customColors={{
-          resting: colors.interactive.warning__resting.rgba,
-          hover: colors.interactive.warning__hover.rgba,
-          backgroundHover: colors.ui.background__warning.rgba,
-        }}
-      />
-    </>
-  );
-};
-
 export default MenuBar;
